Added the routes to add a new sneaker or a new admin
The form to add a new sneaker has the option to upload a file, at the moment the proyect save the images local, for the second release this will be online.

I have to add views and routes to edit and delete an admin or sneaker.
The sneaker edit view can upload a new size for the same sneaker.